refactor(#30): move /services/*DTO.java to /dtos
