+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/controllers/users/RegisterUserController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.zerowaste.zerowaste.services.users.RegisterUserDTO;
+import com.zerowaste.zerowaste.dtos.RegisterUserDTO;
 import com.zerowaste.zerowaste.services.users.RegisterUserService;
 import com.zerowaste.zerowaste.services.users.exceptions.UserWithSameEmailAlreadyExistsException;
