+++ backend/zerowaste/src/test/java/com/zerowaste/zerowaste/application/services/users/RegisterUserServiceTest.java
@@ -17,10 +17,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.util.ReflectionTestUtils;
 
+import com.zerowaste.zerowaste.dtos.RegisterUserDTO;
 import com.zerowaste.zerowaste.models.user.User;
 import com.zerowaste.zerowaste.models.user.UserRole;
 import com.zerowaste.zerowaste.repositories.UsersRepository;
-import com.zerowaste.zerowaste.services.users.RegisterUserDTO;
 import com.zerowaste.zerowaste.services.users.RegisterUserService;
 import com.zerowaste.zerowaste.services.users.exceptions.UserWithSameEmailAlreadyExistsException;
